caravan_project/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose[cryptography]
httpx
email-validator
passlib
bcrypt==3.2.2
python-multipart
